移除非Linux平台的GPIO提供者实现，修复设备缓存中的方法调用错误，更新PylonTech点位定义，新增`ValueType`字段以提升点位类型描述的准确性和一致性。

+++ application/api/device/v1/device_history.go
@@ -2,6 +2,7 @@ package v1
 
 import (
 	"common/c_chart"
+
 	"github.com/gogf/gf/v2/frame/g"
 )
 
@@ -9,8 +10,8 @@ type PostDeviceHistoryReq struct {
 	g.Meta        `path:"/device/history" method:"post" tags:"设备相关" summary:"获取设备历史数据"`
 	DeviceId      string   `json:"deviceId" v:"required" dc:"设备ID"`
 	TelemetryKeys []string `json:"telemetryKeys" v:"required" dc:"遥测点位名称列表"`
-	StartTime     *int     `json:"startTime" dc:"开始时间"`
-	EndTime       *int     `json:"endTime" dc:"结束时间"`
+	StartTime     *int64   `json:"startTime" dc:"开始时间"`
+	EndTime       *int64   `json:"endTime" dc:"结束时间"`
 	Step          int      `json:"step" dc:"步长"`
 }
